refactor: split BVH into multiple modules
